requirements.txt
pandas==1.1.5
spacy==2.2.4
scikit-learn==0.22.2.post1
Pillow==8.2.0
pytesseract==0.3.8
pdf2image
streamlit==0.82.0



